+++ android/engine/src/main/java/org/smartregister/fhircore/engine/ui/login/LoginActivity.kt
@@ -72,7 +72,6 @@ class LoginActivity :
         }
       }
       launchDialPad.observe(this@LoginActivity) { if (!it.isNullOrEmpty()) launchDialPad(it) }
-      appLogoResourceFile = getApplicationConfiguration().appLogoIconResourceFile
     }
 
     if (configurationRegistry.isAppIdInitialized()) {