+++ android/engine/src/main/java/org/smartregister/fhircore/engine/ui/login/LoginScreen.kt
@@ -59,8 +59,6 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.asImageBitmap
-import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
@@ -73,7 +71,6 @@ import androidx.compose.ui.text.style.TextDecoration
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import androidx.core.graphics.drawable.toBitmap
 import org.smartregister.fhircore.engine.R
 import org.smartregister.fhircore.engine.configuration.view.LoginViewConfiguration
 import org.smartregister.fhircore.engine.configuration.view.loginViewConfigurationOf
@@ -83,7 +80,6 @@ import org.smartregister.fhircore.engine.ui.theme.LoginButtonColor
 import org.smartregister.fhircore.engine.ui.theme.LoginDarkColor
 import org.smartregister.fhircore.engine.ui.theme.LoginFieldBackgroundColor
 import org.smartregister.fhircore.engine.util.annotation.ExcludeFromJacocoGeneratedReport
-import org.smartregister.fhircore.engine.util.extension.getDrawable
 
 const val APP_NAME_TEXT_TAG = "aapNameTextTag"
 const val USERNAME_FIELD_TAG = "usernameFieldTag"
@@ -114,7 +110,6 @@ fun LoginScreen(loginViewModel: LoginViewModel) {
     onLoginButtonClicked = { loginViewModel.attemptRemoteLogin() },
     loginError = loginError,
     showProgressBar = showProgressBar,
-    appLogoResFile = loginViewModel.appLogoResourceFile
   )
 }
 
@@ -130,7 +125,6 @@ fun LoginPage(
   modifier: Modifier = Modifier,
   loginError: String = "",
   showProgressBar: Boolean = false,
-  appLogoResFile: String = ""
 ) {
   var showPassword by remember { mutableStateOf(false) }
   val backgroundColor = if (viewConfiguration.darkMode) LoginBackgroundColor else Color.White
@@ -163,7 +157,7 @@ fun LoginPage(
         // TODO Add configurable logo. Images to be downloaded from server
         if (viewConfiguration.showLogo) {
           Image(
-            bitmap = LocalContext.current.getDrawable(appLogoResFile).toBitmap().asImageBitmap(),
+            painter = painterResource(R.drawable.ic_app_logo),
             contentDescription = stringResource(id = R.string.app_logo),
             modifier =
               modifier
@@ -377,8 +371,7 @@ fun LoginScreenPreview() {
     password = "",
     onPasswordChanged = {},
     forgotPassword = {},
-    onLoginButtonClicked = {},
-    appLogoResFile = "ic_launcher"
+    onLoginButtonClicked = {}
   )
 }
 
@@ -393,7 +386,6 @@ fun LoginScreenPreviewDarkMode() {
     password = "",
     onPasswordChanged = {},
     forgotPassword = {},
-    onLoginButtonClicked = {},
-    appLogoResFile = "ic_launcher"
+    onLoginButtonClicked = {}
   )
 }