+++ CHANGELOG.md
@@ -51,6 +51,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Engine | Fixed login authentication issue.
 - Engine | Implement language switcher on profile page
 - Engine | Add tasking abstraction
+- Engine | Integrate Group resource for family representation
 
 ### Fixed
 - ANC | Resolved a bug on measure reporting : App crash when All patient selected